rename evm to rust-evm
